requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt
chainlit==2.1.0
cohere==4.37
openai==1.58.1
python-dotenv==1.0.0
pydantic==2.10.1
numpy
langchain 
langchain_openai 
langchain-community 
langgraph 
arxiv 
langchain-google-community 
googlesearch-python
websockets
